+++ CHANGELOG.md
@@ -185,6 +185,8 @@ there.
   issues (can be disabled in the preferences)
 - **New Feature**: In the combined filemanager mode, you can now see writing
   targets as a ring progress bar in the position of the file icon
+- **New Feature**: Zettlr can now automatically update internal links (Wiki/
+  Zettelkasten style) if you rename a file
 - The TableEditor now has a clear data saving strategy: Whenever you change a
   table, you need to manually click the disk icon at the top of the table to
   apply all your changes to the document so that it can then be saved